# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "servers": [],
  "info": {
    "title": "FutHub API",
    "version": "1.0",
    "description": "API do FutHub para jogo, automações Discord e administração."
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/v1/cards": {
      "get": {
        "tags": ["Mercado de cards"],
        "operationId": "listCardMarket",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": true
          },
          {
            "name": "positions",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["CA", "GOL", "LD", "LE", "MA", "MC", "PD", "PE", "VOL", "ZAG"]
              }
            },
            "required": false
          },
          {
            "name": "minOverall",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": true
          },
          {
            "name": "maxOverall",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": true
          },
          {
            "name": "teamId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "collectionId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["name", "overall", "recent"]
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "items": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "imageUrl": {
                            "type": "string"
                          },
                          "overall": {
                            "type": "number"
                          },
                          "position": {
                            "type": "string",
                            "enum": ["CA", "GOL", "LD", "LE", "MA", "MC", "PD", "PE", "VOL", "ZAG"]
                          },
                          "secondaryPositions": {
                            "items": {
                              "type": "string",
                              "enum": [
                                "CA",
                                "GOL",
                                "LD",
                                "LE",
                                "MA",
                                "MC",
                                "PD",
                                "PE",
                                "VOL",
                                "ZAG"
                              ]
                            },
                            "type": "array"
                          },
                          "defense": {
                            "type": "number"
                          },
                          "attack": {
                            "type": "number"
                          },
                          "creation": {
                            "type": "number"
                          },
                          "passing": {
                            "type": "number"
                          },
                          "control": {
                            "type": "number"
                          },
                          "marking": {
                            "type": "number"
                          },
                          "pace": {
                            "type": "number"
                          },
                          "dribbling": {
                            "type": "number"
                          },
                          "finishing": {
                            "type": "number"
                          },
                          "price": {
                            "type": "number"
                          },
                          "team": {
                            "additionalProperties": false,
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "emoji": {
                                "type": "string"
                              }
                            },
                            "required": ["id", "name", "emoji"],
                            "type": "object"
                          },
                          "collection": {
                            "additionalProperties": false,
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "emoji": {
                                "type": "string"
                              }
                            },
                            "required": ["id", "name", "emoji"],
                            "type": "object"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "imageUrl",
                          "overall",
                          "position",
                          "secondaryPositions",
                          "defense",
                          "attack",
                          "creation",
                          "passing",
                          "control",
                          "marking",
                          "pace",
                          "dribbling",
                          "finishing",
                          "price",
                          "team",
                          "collection"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "pageSize": {
                      "type": "number",
                      "enum": [10]
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  },
                  "required": ["items", "total", "page", "pageSize", "totalPages"],
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": ["v1", "cards", "listCards"],
        "x-samchon-controller": "CardMarketController"
      }
    },
    "/v1/cards/catalog": {
      "get": {
        "tags": ["Mercado de cards"],
        "operationId": "listCardMarketCatalog",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "teams": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "emoji": {
                            "type": "string"
                          }
                        },
                        "required": ["id", "name", "emoji"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "collections": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "emoji": {
                            "type": "string"
                          }
                        },
                        "required": ["id", "name", "emoji"],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": ["teams", "collections"],
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": ["v1", "cards", "catalog", "listCatalog"],
        "x-samchon-controller": "CardMarketController"
      }
    },
    "/v1/cards/{cardId}/purchase": {
      "post": {
        "tags": ["Mercado de cards"],
        "operationId": "purchaseCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "identity": {
                    "additionalProperties": false,
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "avatarUrl": {
                        "type": "string",
                        "nullable": true
                      }
                    },
                    "required": ["id", "name", "avatarUrl"],
                    "type": "object"
                  }
                },
                "required": ["identity"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "userCardId": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "number"
                    },
                    "price": {
                      "type": "number"
                    }
                  },
                  "required": ["userCardId", "balance", "price"],
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": ["v1", "cards", "purchase", "purchaseCard"],
        "x-samchon-controller": "CardMarketController"
      }
    },
    "/v1/cards/sell": {
      "post": {
        "tags": ["Mercado de cards"],
        "operationId": "sellCards",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "identity": {
                    "additionalProperties": false,
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "avatarUrl": {
                        "type": "string",
                        "nullable": true
                      }
                    },
                    "required": ["id", "name", "avatarUrl"],
                    "type": "object"
                  },
                  "userCardIds": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "required": ["identity", "userCardIds"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "userCardIds": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "balance": {
                      "type": "number"
                    },
                    "amount": {
                      "type": "number"
                    }
                  },
                  "required": ["userCardIds", "balance", "amount"],
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": ["v1", "cards", "sell", "sellCards"],
        "x-samchon-controller": "CardMarketController"
      }
    },
...
  },
...
}